requirements.txt
flask>=2.3.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.6
ConfigParser==5.3.0
html2image>=2.0.0
moment>=0.11.2

# Compatibilidad entre pysftp y paramiko
pysftp==0.2.9
paramiko==2.12.0
